Move Compliance Officer maintenance pages
